requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.5
django-cleanup==5.1.0
djangorestframework==3.12.2
httpie==2.3.0
idna==2.10
Pillow==8.1.0
Pygments==2.7.4
PySocks==1.7.1
pytz==2020.5
requests==2.25.1
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.2
